langchain_project/pyproject.toml
[project]
name = "langchain-project"
version = "0.1.0"
description = ""
authors = [
    {name = "yoerl",email = "jokerbook365@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "langchain==0.2.16",
    "tiktoken==0.7.0",
    "langchain-community==0.2.16",
    "seaborn==0.13.2",
    "langchain-chroma==0.1.3",
    "langchain-openai==0.1.23",
    "gradio==4.44.0",
    "pypdf==4.3.1",
    "bs4==0.0.2",
    "jq==1.8.0",
    "transformers==4.44.2",
    "langchain-experimental==0.0.65",
    "langchain-huggingface==0.0.3",
    "sentence-transformers==3.1.0",
    "langchain-ollama==0.1.3",
    "deepl==1.18.0",
    "langdetect==1.0.9",
    "faiss-cpu==1.8.0.post1",
    "rank-bm25==0.2.2",
    "kiwipiepy==0.18.1",
    "openpyxl==3.1.5",
    "krag==0.0.25",
    "langchain-anthropic==0.1.23",
    "langchain-google-genai==1.0.10",
    "langchain-groq==0.1.10",
    "python-dotenv==1.0.1",
    "fastapi==0.110.0",
    "uvicorn==0.27.1",
    "pydantic>=2.7.4,<3.0.0"
]

[tool.poetry]
packages = [{include = "langchain_project", from = "src"}]


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
